Changes suggested by Cyenyxe
