+++ src/main/java/uk/ac/ebi/ampt2d/accessioning/commons/core/BasicMonotonicAccessioningService.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  *
  */
-package uk.ac.ebi.ampt2d.accessioning.commons.accessioning;
+package uk.ac.ebi.ampt2d.accessioning.commons.core;
 
 import org.springframework.beans.factory.InitializingBean;
 import uk.ac.ebi.ampt2d.accessioning.commons.generators.monotonic.MonotonicAccessionGenerator;
