+++ src/main/java/uk/ac/ebi/ampt2d/accessioning/commons/generators/monotonic/MonotonicAccessionGenerator.java
@@ -17,7 +17,7 @@
  */
 package uk.ac.ebi.ampt2d.accessioning.commons.generators.monotonic;
 
-import uk.ac.ebi.ampt2d.accessioning.commons.accessioning.SaveResponse;
+import uk.ac.ebi.ampt2d.accessioning.commons.core.SaveResponse;
 import uk.ac.ebi.ampt2d.accessioning.commons.generators.AccessionGenerator;
 import uk.ac.ebi.ampt2d.accessioning.commons.generators.ModelHashAccession;
 import uk.ac.ebi.ampt2d.accessioning.commons.generators.exceptions.AccessionCouldNotBeGeneratedException;
