+++ src/test/java/uk/ac/ebi/ampt2d/accessioning/commons/core/BasicAccessioningServiceSaveDelegateTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  *
  */
-package uk.ac.ebi.ampt2d.accessioning.commons.accessioning;
+package uk.ac.ebi.ampt2d.accessioning.commons.core;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
