+++ src/test/java/uk/ac/ebi/ampt2d/accessioning/commons/generators/monotonic/MonotonicRangeTest.java
@@ -19,6 +19,7 @@ package uk.ac.ebi.ampt2d.accessioning.commons.generators.monotonic;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.junit.runners.BlockJUnit4ClassRunner;
 import org.springframework.test.context.junit4.SpringRunner;
 
 import java.util.Arrays;
@@ -28,7 +29,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-@RunWith(SpringRunner.class)
+@RunWith(BlockJUnit4ClassRunner.class)
 public class MonotonicRangeTest {
 
     @Test
